requirements.txt
numpy==1.24.4
transformers==4.51.3
diffusers==0.31.0
Pillow==9.5.0
opencv-python==4.10.0.84
segment-anything-py==1.0.1
huggingface-hub==0.31.2
modelscope==1.25.0
invisible-watermark==0.2.0
kornia==0.8.1
datasets==2.21.0
addict==2.4.0
sortedcontainers==2.4.0
simplejson==3.20.1
sentencepiece==0.2.0

--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchvision==0.21.0
